Create StatReset.js
Lets the GM reset stats of every selected token. This both resets the values on the character sheet connected to the token and it resets the bars (that may or may not be linked) on the token. Most values are reset to their maximums, but Fatigue is reset to 0. This code assumes that the green bar (bar1) is Fatigue, the blue bar (bar2) is Fate, and the red bar (bar3) is Wounds. bar2 and bar3 will be reset to their maximums. bar1 will be reset to 0 as it represents Fatigue.