Create CharacterImporter.js
Very temperamental code! Be warned!

Allows the GM to copy paste the text of a character from a Rogue Heresy book into a character sheet. This code will then read the text and attempt to decipher all the Attributes of the character, all of the attacks of the character (making Macros for them), and creating links to any Skills, Talents, Traits, or Gear that they have (that are already in your Handout Library).