Fate Cleanup
Used a surer method to detect when the user is a GM
playerIsGM(playerid)

I removed some old testing features from the code that were no longer
needed.

Used _.each for both !Fate and !FateReset. See below for reasons.

!Fate: Though it is unlikely that a player will have multiple characters
and attempt to spend a Fate Point for each of them, it is better to
allow for this then for them to unknowingly select multiple characters
and keep spending a Fate Point on the wrong character (whichever one the
API decides is character[0]).

!FateReset: Previously the code searched through every single character
in the library and reset their Fate Points. My library has grown quite
large and this function became too inconvenient to use. By only reseting
the Fate Points for those characters selected, the processing time is
dramatically reduced.